lambda/powerpoint_generator/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
python-pptx>=0.6.22